requirements.txt
# Python 3.12 compatible versions
# Core dependencies
numpy>=1.26.0  # Python 3.12 support from 1.26
pandas>=2.1.0  # Python 3.12 support from 2.1
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Machine Learning
scikit-learn>=1.3.2  # Full Python 3.12 support
xgboost>=2.0.0
lightgbm>=4.1.0

# Deep Learning
torch>=2.1.0  # Python 3.12 support
torchvision>=0.16.0
tensorflow>=2.15.0  # Python 3.12 support

# NLP and LLMs
transformers>=4.36.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
openai>=1.0.0
anthropic>=0.8.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# AI Agents
autogen>=0.2.0
chromadb>=0.4.0

# Jupyter and notebooks
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.1.0
notebook>=7.0.0
nbformat>=5.9.0
ipykernel>=6.26.0

# Web apps
streamlit>=1.29.0
gradio>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
tqdm>=4.66.0
PyYAML>=6.0

# Data handling
openpyxl>=3.1.0
Pillow>=10.0.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
